3) Disease Text Model/requirements.txt
pytorch-tabnet
scikit-learn
pandas
numpy
streamlit
matplotlib
seaborn
pillow
joblib
torch